requirements.txt
numpy
Pillow
SimpleITK
scikit-image
monai
einops
hausdorff
pandas
batchgenerators
opencv-python
easydict
ipykernel
matplotlib
tensorboard
numba
seaborn